backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sse-starlette==2.1.0
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.3.1

# LangGraph + LangChain
langgraph>=0.1.19
langchain>=0.2.6
langchain-core>=0.2.22
langchain-groq==0.1.6
langchain-openai==0.1.14
langchain-pinecone==0.1.1
langchain-community>=0.2.6

# Groq
groq==0.9.0

# Pinecone
pinecone-client>=3.2.2

# MongoDB
motor==3.4.0
pymongo==4.7.3

# Redis / Valkey
redis==5.0.6

# MinIO / S3
boto3==1.34.144

# Export
reportlab==4.2.2
python-docx==1.1.2

# Utilities
httpx==0.27.0
python-multipart==0.0.9
uuid==1.30
